+++ annotations/nocedit.exe/pseudocode/src/core/cloth.cpp/CCloth_showMenu_FUN_0043ddf0.cpp
@@ -43,7 +43,7 @@ void __cdecl core_cloth_cpp_CCloth_showMenu_FUN_0043ddf0(CCloth *this_ptr)
     case 0x31:
       iVar4 = shape_edittool_cpp_CEditorTools_showFileSelectionDialog_FUN_0049f270
                         (g_CEditorToolsPtr,"Load cloth file","models",
-                         "*.cth",(int)local_178,0);
+                         "*.cth",local_178,0);
       if (iVar4 != 0) {
         core_cloth_cpp_CCloth_load_FUN_00438cf0(this_ptr,local_178);
       }
@@ -62,7 +62,7 @@ void __cdecl core_cloth_cpp_CCloth_showMenu_FUN_0043ddf0(CCloth *this_ptr)
     case 0x34:
       iVar4 = shape_edittool_cpp_CEditorTools_showFileSelectionDialog_FUN_0049f270
                         (g_CEditorToolsPtr,"Select deformable model","models",
-                         "*.dfm",(int)local_278,0);
+                         "*.dfm",local_278,0);
       if (iVar4 != 0) {
         core_skeleton_cpp_CDeformableModelInstance_init_FUN_005a0840
                   (&g_CDeformableModelInstanceInstance,local_278);
@@ -83,7 +83,7 @@ void __cdecl core_cloth_cpp_CCloth_showMenu_FUN_0043ddf0(CCloth *this_ptr)
     case 0x37:
       iVar3 = shape_edittool_cpp_CEditorTools_showFileSelectionDialog_FUN_0049f270
                         (g_CEditorToolsPtr,"Create cloth file from KFM","models",
-                         "*.kfm",(int)local_78,0);
+                         "*.kfm",local_78,0);
       if (iVar3 != 0) {
         splitpath(local_78,(char *)0x0,(char *)0x0,local_178,(char *)0x0);
         pcVar4 = ".cth";
