+++ annotations/nocedit.exe/pseudocode/src/core/game.cpp/CGame_showChapterSelect_FUN_004e1cb0.cpp
@@ -86,7 +86,7 @@ void __cdecl core_game_cpp_CGame_showChapterSelect_FUN_004e1cb0(CGame *this_ptr,
   if (bVar8 || p_Var3 != (_FILE *)0x0) {
     iVar4 = shape_edittool_cpp_CEditorTools_showFileSelectionDialog_FUN_0049f270
                       (g_CEditorToolsPtr,"Select mission to play","world",
-                       "*.msn",(int)local_144,0);
+                       "*.msn",local_144,0);
     if (iVar4 == 0) {
       g_EditorFont = local_1c;
       return;
